lambda/db-init/requirements.txt
psycopg2-binary==2.9.10
cfnresponse==1.1.2